% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{bspm-options}
\alias{bspm-options}
\title{Package Options}
\description{
List of \code{\link{options}()} supported to configure \pkg{bspm}'s behavior.
In general, these should be set \emph{before} calling any package function.
}
\section{Options specific to \pkg{bspm}}{

\describe{
\item{\code{bspm.backend.check}:}{logical, default \code{TRUE}. If false,
  the initial check on \code{\link{enable}()} is not performed.}
\item{\code{bspm.version.check}:}{logical, default \code{TRUE}. If false,
  as many binaries are installed as possible without any version check,
  and then installation from source is used as a fallback.}
\item{\code{bspm.sudo.autodetect}:}{logical, default \code{FALSE}. If true,
  enables autodetection and selection of password-less \code{sudo}.}
\item{\code{bspm.sudo}:}{logical, default \code{FALSE}. If true, forces
  \code{sudo} unconditionally as the preferred mechanism.}
}
}

\section{Options from base \R}{

These are used in the same way as in base \R. See \code{\link{options}} for
a detailed description.
\itemize{
\item{\code{askYesNo}}
\item{\code{install.packages.compile.from.source}}
}
}

