% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\docType{package}
\name{bspm-package}
\alias{bspm-package}
\title{\pkg{bspm}: Bridge to System Package Manager}
\description{
Enables binary package installations on Linux distributions.
Provides functions to manage packages via the distribution's package
manager. Also provides transparent integration with R's \code{install.packages}
and a fallback mechanism. When installed as a system package, interacts
with the system's package manager without requiring administrative
privileges via an integrated D-Bus service; otherwise, uses sudo.
Currently, the following backends are supported: DNF, APT.
}
\references{
\url{https://github.com/Enchufa2/bspm}
}
\seealso{
\code{\link{manager}}, \code{\link{integration}}
}
\author{
Iñaki Ucar
}
