% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scripts.R
\name{bspm-scripts}
\alias{bspm-scripts}
\title{Call Internal Scripts}
\description{
Internal scripts can be called via \code{Rscript} (see examples).
}
\examples{
\dontrun{
# get a list of available scripts with descriptions
Rscript -e bspm:::scripts

# see a script's help
Rscript -e bspm:::scripts <script_name> -h

# run a script
Rscript -e bspm:::scripts <script_name> [args]
}

}
