% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{discover}
\alias{discover}
\title{Discover Packages from System Repositories}
\usage{
discover()
}
\description{
Talk to the system package manager to discover new packages. Needed only
when e.g. a new repository is added that contains packages with different
prefixes (for example, your system repositories may provide packages called
\code{r-cran-*} and \code{r-bioc-*} and then you add a new repository that
provides packages called \code{r-github-*}).
}
