% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integration.R
\name{enable}
\alias{enable}
\alias{disable}
\title{Enable/Disable Bridge to System Package Manager}
\usage{
enable()

disable()
}
\description{
Functions to enable or disable the integration of \code{\link{install_sys}}
into \code{\link{install.packages}}. When enabled, packages are installed
transparently from system repositories if available, and from CRAN if not.
}
