% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manager.R
\name{install_sys}
\alias{install_sys}
\alias{remove_sys}
\title{Install Binary Packages from System Repositories}
\usage{
install_sys(pkgs)

remove_sys(pkgs)
}
\arguments{
\item{pkgs}{character vector of CRAN names of packages.}
}
\value{
Invisibly, a character vector of the names of packages not available.
}
\description{
Talk to the system package manager to download and install or remove
packages from system repositories.
}
\details{
The root user talks directly to the system package manager.
Non-root users talk to the accompanying D-Bus service, which performs the
required actions and returns packages that could not be processed.
}
