% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{sysreqs}
\alias{sysreqs}
\title{Find System Requirements}
\usage{
sysreqs(pkgs, ..., type = c("build", "run"), distro = NULL,
  collapse = FALSE)
}
\arguments{
\item{pkgs}{character vector of R package names.}

\item{...}{unused, reserved for future expansion.}

\item{type}{type of requirements: build- or run-time requirements, or both.}

\item{distro}{name of the Linux distribution; if nothing is provided, the
function tries to detect it automatically.}

\item{collapse}{whether to collapse the requirements into a single line.}
}
\value{
A character vector of system requirements.
}
\description{
Compute the system requirements (system libraries; operating system packages)
required by a set of R packages.
}
\details{
This function relies on https://github.com/cran4linux/sysreqs.
}
